fix(scheduler/coreos): have announce/log sidecars wait for app
Currently, the announce and log containers can timeout on startup if
the app takes longer than the default systemd timeout to come up.

Instead, we now use systemd's After= to make sure the unit doesn't
try to come up until the app is running.

This alleviates an issue where we see the app unit running, but the
announce and logger units in a failed state.

TESTING: rebuild and restart the controller, then scale apps:
```console
	$ make -C controller/ build restart
```